requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pillow>=9.0.0

# Vision model support
timm>=0.9.0
clip-by-openai>=1.0
open-clip-torch>=2.20.0

# Language models
accelerate>=0.20.0
tokenizers>=0.13.0

# Configuration and utilities
pydantic>=2.0.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0
loguru>=0.7.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Optional: Fast implementations
# uvloop>=0.17.0  # For async performance
# orjson>=3.8.0   # For fast JSON processing

# Robot communication (optional)
# pyserial>=3.5   # For serial communication
# websockets>=11.0  # For WebSocket communication
# grpcio>=1.50.0  # For gRPC communication 